isourceit-server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
cssselect2==0.7.0
dnspython==2.3.0
eventlet==0.33.3
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-SocketIO==5.3.3
frozenlist==1.3.3
greenlet==2.0.2
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
oscrypto==1.3.0
Pillow==9.5.0
pycparser==2.21
pydantic==1.10.7
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pymongo==4.3.3
pypdf==3.8.1
pypng==0.20220715.0
python-bidi==0.4.2
python-engineio==4.4.1
python-socketio==5.8.0
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
redis==4.5.4
reportlab==3.6.13
requests==2.29.0
six==1.16.0
sseclient-py==1.7.2
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.3
xhtml2pdf==0.2.10
yarl==1.9.2
